Clean last PR in css, remove footer links.
We have removed some dependencies from the project, such as the week number by RobG, minimal.js and textlayerbuilder.js. They are no longer needed to be listed in the footer.

I also commented out some stuff in the css to fix a merge conflict. That comment is now removed.